+++ eosws/action_test.go
@@ -213,6 +213,7 @@ func TestOnGetActionsTraces(t *testing.T) {
 				nil,
 				NewTestIrreversibleFinder("00000001a", nil),
 				0,
+				12,
 			)
 
 			conn, closer := newTestConnection(t, handler, &testCredentials{startBlock: 2})