+++ eosws/headinfo_test.go
@@ -63,6 +63,7 @@ func Test_onGetHeadInfo(t *testing.T) {
 				nil,
 				NewTestIrreversibleFinder("00000002a", nil),
 				0,
+				12,
 			)
 
 			conn, closer := newTestConnection(t, handler)